+++ src/test/model/inline-document-descriptor_test.ts
@@ -21,8 +21,11 @@ import {resolvedUrl} from '../test-utils';
 suite('correctSourceRange', function() {
   test('handles undefined gracefully', function() {
     const zeroPosition = {line: 0, column: 0};
-    const zeroSourceRange:
-        SourceRange = {file: resolvedUrl`foo`, start: zeroPosition, end: zeroPosition};
+    const zeroSourceRange: SourceRange = {
+      file: resolvedUrl`foo`,
+      start: zeroPosition,
+      end: zeroPosition
+    };
     const zeroLocationOffset: LocationOffset = {line: 0, col: 0};
 
     assert.equal(correctSourceRange(undefined, undefined), undefined);